requirements.txt
absl-py==0.6.1
alembic==1.0.1
aniso8601==3.0.2
astor==0.7.1
astroid==2.0.4
bcrypt==3.1.4
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-GraphQL==2.0.0
Flask-Migrate==2.3.0
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gast==0.2.0
graphene==2.1.3
graphene-sqlalchemy==2.1.0
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
grpcio==1.16.0
grpcio-tools==1.16.0
h5py==2.8.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.3.1
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.15.3
pkg-resources==0.0.0
promise==2.2.1
protobuf==3.6.1
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.6
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.12
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
